Simplify some parts of the custom forms code based on new understanding
Due to some inconsistencies with how the new expander implemented
internal definition contexts, some of the code in this blog post was
more complex than it needed to be. Now that
racket/racket@3531cb24f4ca181927133cf33a1245c5660678d8 brings the new
expander in line with the old one, some of the code can be simplified.